Create Run policy trigger if publicIP match.sh
This is an enhancement of the Run Ppolicy of network connection is off campus script.
This is generalized to either match condition TRUE|FALSE for on or off the publicIP addresses provided in parameter 5.  Policy triggers to be provided in parameter 6.  Parameter 5 and 6 handle either single value or list of space separated IPv4 addresses.